% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihs_cache.R
\name{ihs_cache_clear}
\alias{ihs_cache_clear}
\title{Clear Cached IHS Data}
\usage{
ihs_cache_clear(round = NULL)
}
\arguments{
\item{round}{A specific round to clear (e.g. \code{"IHS5"}). If \code{NULL}, asks for confirmation
to clear all IHS data depending on the interactivity of the session. Defaults to \code{NULL}.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Removes downloaded datasets from the internal package cache. This is useful for freeing
up disk space. You can clear the cache for specific rounds or entirely.
}
\examples{
\dontrun{
# Clear all
ihs_cache_clear()

# Clear only IHS3 data
ihs_cache_clear(round = "IHS3")
}
}
