% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihs_cache.R
\name{ihs_cache_info}
\alias{ihs_cache_info}
\title{Display Information About Cached IHS Data}
\usage{
ihs_cache_info()
}
\value{
A \code{tibble} summarizing the cached files.
}
\description{
Scans the internal \code{ihsMW} cache directory and reports on any previously
downloaded datasets.
}
\examples{
\dontrun{
ihs_cache_info()
}
}
