% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihs_auth.R
\name{ihs_key_set}
\alias{ihs_key_set}
\title{Set up World Bank Microdata API Key (Alias)}
\usage{
ihs_key_set(key)
}
\arguments{
\item{key}{A single string containing your World Bank Microdata API key.}
}
\value{
Invisibly returns the API key.
}
\description{
A wrapper for \code{ihs_auth()} meant for use in scripted or non-interactive environments.
}
\examples{
\dontrun{
ihs_key_set("paste_your_key_here")
}
}
