% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihs_label.R
\name{ihs_label}
\alias{ihs_label}
\title{Fetch specific variable label locally or remotely}
\usage{
ihs_label(variable, round = "IHS5")
}
\arguments{
\item{variable}{A single character variable map or harmonised standard to inspect precisely.}

\item{round}{The physical survey round to tie it structurally to if verifying non-harmonised entries. Default \code{"IHS5"}.}
}
\value{
The extracted label mapping directly against what the variable corresponds natively to.
}
\description{
Quickly deciphers what an individual variable physically measures. Looks through the offline dataset initially via harmonised mappings and seamlessly falls through NADA otherwise.
}
\examples{
ihs_label("rexp_cat01")
}
