% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihs_label.R
\name{ihs_modules}
\alias{ihs_modules}
\title{Inspect available modules for a study}
\usage{
ihs_modules(round = "IHS5")
}
\arguments{
\item{round}{A specific round to fetch dataset structures safely scoped against. (e.g. \code{"IHS5"}).}
}
\value{
Invisibly returns a tibble mapping underlying module variables mapping locally.
}
\description{
Profiles the file hierarchy explicitly for each data survey pulling nested variables counts efficiently per dataset.
}
\examples{
\dontrun{
ihs_modules("IHS5")
}
}
