% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihs_clean.R
\name{ihs_standardize_missing}
\alias{ihs_standardize_missing}
\title{Standardize Survey Missing Codes}
\usage{
ihs_standardize_missing(data)
}
\arguments{
\item{data}{A data.frame}
}
\value{
A data.frame with missing values standardized
}
\description{
Converts common negative missing codes (like -99 for "Refused" or -98 for "Don't Know") 
into standard R `NA` values to prevent them from skewing numeric calculations.
}
